requirements.txt
anyio==4.12.1
async-timeout==5.0.1
certifi==2026.2.25
charset-normalizer==3.4.5
elastic-transport==9.2.1
elasticsearch==9.3.0
exceptiongroup==1.3.1
idna==3.11
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
redis==7.3.0
requests==2.32.5
six==1.17.0
sniffio==1.3.1
typing_extensions==4.15.0
urllib3==2.6.3
